[iOS] Remove ConfirmationAlertViewController gradient mask offset
The gradient mask of ConfirmationAlertViewController's scroll view
was getting offset when it should simply have taken the view's
size (the comment was correct, but the code was not). If the scroll
view was in its origin position, the gradient mask was properly set
up, but when the scroll view appeared while already being scrolled,
the gradient mask would cut off a portion of the scroll view.

The fix is to not copy the scroll view's offset to the gradient mask's
offset, as the gradient mask is relative to the scroll view.

Fixed: 447384479
Change-Id: I9578373be704a01265d42e1f3332c2c8d777dcc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990591
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522202}