[iOS] Fix the cell Content View for accessibility
When enabling accessibility font size, the layout should be fully
vertical.
This CL also removes the max preferred max width for label as it
wasn't working with the accessibility as the width used was the
one from the content view and not the cell.

https://screenshot.googleplex.com/AABSgNk3N2vbX3V
https://screenshot.googleplex.com/6mzDmg4amSJrsyU

Bug: 443034511
Change-Id: I69dc2befdcb232d0d47599e833b9d0771b1a8bbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6988735
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1522208}