Don't show update prompts if the data is already saved
See bug description of crbug.com/447295932 for a description of the
relevant scenario. The current implementation and tests expect that in
this case an update prompt is shown. This CL changes the behavior and
instead doesn't show a prompt (or potentially shows a migration prompt).

As noted in crbug.com/447295932#comment3, it is true that this is
ultimately a deduplication issue and our deduplication logic should
simply run more frequently. However, given the recent refactoring, it
is simple to change the import logic to accommodate this case and doing
so makes several lengthy existing tests obsolete. So even if we fix the
deduplication logic (which we should), this CL seems like a win.
Moreover, changing the deduplication logic is probably much trickier,
since it is significantly more expensive in terms of runtime. So this
CL is at least a simple way to stop the bleeding.

Notably, with this change it becomes impossible for an update prompt and
silent updates to happen as part of the same form submission. If a
user-visible update for A and a silent update for B are possible, A is
a subset of B.

Fixed: 447295932
Change-Id: I8bdb3fcb682730b03a0f00f8f98bbc12677cb2d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6994558
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1522212}