[AccountName] Clear related metadata prefs on sign out
This CL makes the AccountNameEmailStore follow H/W in how sign outs are
handled.

When the sync transport state is set to disabled is very likely means
that the signout happened. When it is detected, creation of the new
profile is blocked and the existing profile is removed and the prefs
cleared.

Bug: 356845298
Change-Id: I9c6f23dad93ba665bdf75d814ec58eae5a2f4188
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6994282
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Karol Sygiet <sygiet@google.com>
Cr-Commit-Position: refs/heads/main@{#1522224}