[ios] Build ChromeWebView as safe to use for application extensions
As all the infrastructure is in place, it is now possible to build
ChromeWebView as safe to use  for application extension by setting
the parameter app_ext_safe to true.

This remove the need to pass `-fapplication-extension` via configs
as this is now handled by the toolchain internally in the template
ios_framework_bundle(...).

This is part of a multi-CL change:
 -  https://crrev.com/c/6973372
 -  https://crrev.com/c/6973563
 -  https://crrev.com/c/6973405
 -  https://crrev.com/c/6973680
 -  https://crrev.com/i/8609736
 -  https://crrev.com/c/6973583
 => https://crrev.com/c/6973781

Bug: 40226740
Change-Id: I1abbcbc96147e7699d40dec10cf249503edf64ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6973781
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Hiroshi Ichikawa <ichikawa@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522234}