[iOS] Move and rename WebClient::OverrideOpenPanel() to WebStateDelegate
This CL moves OverrideOpenPanel() from the WebClient class to the
WebStateDelegate class so that different implementations of the delegate
can handle requests to run the open panel differently. The method is
also renamed to CanRunOpenPanel() to clearly indicate that
RunOpenPanel() (added in a follow-up change to the same class) cannot be
called if CanRunOpenPanel() returns false.

Additionally the source WebState is now passed as an argument, since
delegate methods should receive the source of the call as argument. This
allows returning a value which depends on the caller.

Bug: 441659098
Change-Id: Id119aecc9b511f4a3423988b146a420b5349a29f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6994541
Auto-Submit: Quentin Pubert <qpubert@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522242}