[FillRecoveryPwd] Report metric upon selecting backup password
Reports PasswordDropdownItemSelected metric if backup password is
selected in KA sheet or TTF. This is the existing metric, which is
reported on all platforms to identify the type of selected item in KA
(on Android) or dropdown (on Desktop).
The CL also introduces a new method and renames to old one to be able to
report PasswordDropdownItemSelected metric separately from other
metrics.

Bug: 422361849
Change-Id: I9bcc78f5a0c6f06e88b5d7b5386336f1ff9ed9fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990131
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522256}