[views-ax] Track nodes in widget tree eligible for serialization
This CL implements tracking of the nodes in the widget tree. This is
necessary to only serialize the accessibility data of nodes that are in
the tree, and trigger updates on the parent node when the view gets
removed.

The solution implemented here starts with two new functions:
* WidgetAXManager::OnChildAdded
* WidgetAXManager::OnChildRemoved

Those functions are called on the parent node when its children change.
We do this because when nodes are added to and removed from the tree, we
must serialize the parent node instead of the added/removed node. When
OnChildAdded or OnChildRemoved is called, we add the parent node to the
pending tree updates vector and serialize it in SendPendingUpdate.

We had to inverse the event and data loops in the SendPendingUpdate
function because it's possible that an event is fired from a child node
that just got added, and unless we serialized the parent node and
therefore added that new node to the tree before we attempted to
serialize the fired event, it would simply be ignored.

Finally, for testing, we added a new mechanism to validate what gets
serialized in WidgetAXManagerTestApi. I refactored other tests that
benefit from this new system too while I was at it.

Bonus: I added a set in which we track which nodes have been serialized
in the ongoing serialization to avoid sending unnecessarily copies of
the same node more than once. This is similar to what is done in Blink,
to improve performance.

Bug: 40672441
Change-Id: Ie1f1a71f304e037df74c96d37743503deb39e13c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6806961
Commit-Queue: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Auto-Submit: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1522270}