[FillRecoveryPwd] Enable feature FillRecoveryPassword by default
The feature enables password recovery flow on Android. As agreed before,
we enable it by default because there are anyway no metrics to analyze
as the usage of APC feature is very small.
Disabled the following tests on Android:
- PasswordRecoveryFlow_AppendsTroubleSigningInSuggestion
- PasswordRecoveryFlow_TroubleSigningInIsAppendedLast
They are valid only for desktop because "Trouble signing in" is not
appended on Android platform.
PasswordRecoveryFlow_AppendsBackupPasswordSuggestion checks that backup
password is appended on Android.

Bug: 414318806
Change-Id: Idbc52614ecf1e51d4b4f14bc8129015eee33cc04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6994400
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522272}