Fix download bubble closing in macOS fullscreen windows.
This changes the downloads BubbleCloser to monitor the top-level
widget the toolbar button is in, rather then the direct widget
parent of the toolbar button. This way even in immersive
fullscreen we end up monitoring the entire browser widget.

Additionally, this also fixes the macOS EventMonitor implementations
to match the Aura implementation, and include events targeting all
children and descendants of the widget being monitored and not
just those events targeting the widget itself.

Bug: 447174550
Change-Id: Iadea937059cd3c71b5b8afc9891092e5cfce4f61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6969166
Reviewed-by: Mark Mentovai <mark@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522305}