[Media App] Perform OCR only when the service is connected.
Media App disconnects from OCR service if accessibility is turned off,
but it might be waiting for an image to send to OCR service, and if this
happens, the image can be sent to an uninitialized OCR service.
To avoid the issue, check if OCR service is connected before sending the
image to it.

Bug: 447820325
Change-Id: I9fc836bfde2f962203efa7a0ebc3cad015abcc0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6992370
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522329}