[BTM] Remove failing CHECK in BtmPageVisitObserver
The CHECK seems to try to enforce a guarantee that isn't provided (the
doc comment for WebContentsObserver::WebAuthnAssertionRequestSucceeded()
doesn't say anything to suggest that notifications will only be issued
for descendant frames of the primary main frame).

The CHECK was added so we would be notified when the assumed guarantee
was violated; it now seems like it's possible that WAA success
notifications are being received "late" (i.e., after primary page
change) and need to be applied to a pending visit. BtmPageVisitObserver
isn't being used for any mitigations at the time of writing, so it's
less important that it correctly applies every WAA success notification;
however, BtmBounceDetector uses the same WCO method, operates under the
same timing assumption, and is actively used for a mitigation, so we may
want to investigate further. This CL adds a TODO to that effect.

[1]:
https://chromium-review.googlesource.com/c/chromium/src/+/6194853/comment/39e474a7_19930792/

Bug: 446194865
Change-Id: Ifb9cd72739d257acf460761585dab8dfc09b5a42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6995079
Reviewed-by: Andrew Liu <liu@chromium.org>
Auto-Submit: Svend Larsen <svend@chromium.org>
Commit-Queue: Andrew Liu <liu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522337}