Actor: Cleanup PageStabilityMonitor navigation interaction
The main functional changes here are:

  * Navigation commit goes straight to invoking a callback
  * Failed navigation re-enters the monitor

Note that the commit signal is only received when RenderDocument is
disabled. Otherwise, a navigation simply replaces a RenderFrame so the
monitor must observe destruction instead in that case. Hence, the
NavigationCommitted state is renamed to RenderFrameGoingAway.

This CL does a fair bit of refactoring in the state machine relating to
navigation states but is mostly adding tests for all these cases.

Fixed: b:436573891
Change-Id: I6f3f68c5404e48826dbf86717c139b6f076d794a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6991969
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522338}