[Signin][Android] Add new seamless sign-in promo layout with two buttons
This CL introduces one of the two new sign-in promo layouts that will be
used in the Seamless Sign-in flow.

This UI is similar to the current one: it has a button to continue the
sign-in process and a button to choose another account.

It can be enabled by enabling the switch ENABLE_SEAMLESS_SIGNIN with the
param seamless-signin-promo-type set to "twoButtons".

Since the new strings for the Seamless sign-in promo layouts aren't
ready yet, this new layout uses the same strings as the current one.

Screenshots in the bug: https://crbug.com/437039646#comment3

Bug: 437039646
Change-Id: I4677e8cbede7c3b1623e80b229bdfdc096c45f31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6973489
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Lucia Giorgi <quacksort@google.com>
Cr-Commit-Position: refs/heads/main@{#1522369}