[SxS] Avoid re-entrency when reversing tabs in split
The logic for reversing tabs in the split assumes the browser widget has
focus. But if you have bluebird open and right click a tab that won't
bring focus to the browser. In that case, activate the multi contents
view before assigning the web contents so focus is restored properly to
the active tab in a split.

Change-Id: I15595661d41ce8ebde901d5de3df583339ba1d4a
Bug: 447369458
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990972
Commit-Queue: Alison Gale <agale@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522371}