[lensoverlay] Fix crash when using zero state CSB.
The Lens overlay controller is making the assumption that the
contextualization controller will always be kOff if the overlay itself
is in state Off. Now that the overlay can be reinvoked, the search
controller needs to make sure the session has not already started in
order to use zero state CSB flow.

Bug: 448127651
Change-Id: Id5ea6d120bf979cd0dc554e6d9956f971e1ac66d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6993803
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1522375}