[contextual_tasks] Add support for getting task context.
This CL introduces the concept of a `ContextualTaskContext`, which is a
representation of the context associated with a `ContextualTask`.

A new method, `GetContextForTask`, has been added to the
`ContextualTasksService` to retrieve this context. The
`ContextualTaskContext` is constructed from a `ContextualTask` and
currently includes URL attachments.

Unit tests for the new functionality have been added.

Bug: b/445841461
Change-Id: Ibc412ca38fceb9260fee6babab254e74ec3b7da9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6990299
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522449}