[WebAuthn] Make GetClientCapabilities accurate for Signal API on Clank
On Android Signal API is not currently available but the capabilities
report shows that it is.

This moves the check out of the renderer process and into the browser.
Desktop Chrome will still always report that it is supported, but
Android will report that it is not.

A forthcoming change that enables Signal API on Android Chrome will
make the capability dependent on the state of the runtime flag.

Fixed: 419292100
Change-Id: I1b9eaa63cff9633a98159b7e6f62a95d33b4d2ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6995918
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522490}