sad_tab: Show OOM sad tab for memory evictions
When a renderer process is terminated with the status
`TERMINATION_STATUS_EVICTED_FOR_MEMORY`, it would previously fall
through to the default case and be treated as a generic crash. This is
misleading, as the termination was a deliberate action to reclaim
memory.

This change maps the eviction status to `SAD_TAB_KIND_OOM`. This ensures
the user is shown the more specific out-of-memory "sad tab" page,
providing better context for the termination.

Bug: 394092280

Change-Id: If77f539e9158c15ab8843dd67e86cb0fb50def5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6996407
Reviewed-by: Tom Lukaszewicz <tluk@chromium.org>
Commit-Queue: Alex Attar <aattar@google.com>
Cr-Commit-Position: refs/heads/main@{#1522498}