[Extensions] Remove globalParams from scripting IDL schema
This entry was added in crrev.com/c/6862906, but due to the nodoc and
nocompile extended attributes it is not actually functional in any way.
To make this API schema more cleanly be able to be translated to WebIDL,
this CL removes this property for now.

Eventually if we do want this to be included in the schema file for
documentation purposes, it would probably be better represented in
WebIDL as an attribute on the main API interface with type object,
although this would require a few small tweaks in the schema compiler to
support this.

Bug: 445495198
Change-Id: I9a540c95a4f075e62ebfeb4fc6072dc9da9d810e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6965950
Commit-Queue: Tim <tjudkins@chromium.org>
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522518}