[AIM] Do not allow reinvoking overlay when feature disabled.
Inadvertently, all entrypoints were allowed to reinvoke the overlay as
they were not marked as disabled due to the side panel state no longer
being accounted for in `IsOverlayActive`. This should only be allowed if
the feature is enabled and there should no change to the visibility of
the Lens entrypoints when the side panel is open.

Bug: 448131721
Change-Id: I42c44de1ad188ee73ad0f063c3d357273c8efce9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6994634
Reviewed-by: Duncan Mercer <mercerd@google.com>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1522519}