[Predictable App Update] Measure uncompressed icon sizes post download
This CL adds a metric to measure the sizes of the icons downloaded as
part of the update process (both old and new). Since less icons will
be downloaded as part of predictable app updating, this can still be
used to measure the effectiveness of the new algorithm, even though
there are some caveats (like less icons downloaded but larger sizes)
etc.

Bug: 447149311
Change-Id: Id93bc82c3db2f586cbce30c3161ed205fd79740f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6993805
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522574}