[clang-tidy] Fixup bugprone-loop-variable-copied-then-modified triggers for expensive-to-copy-types in main Chromium repo, where the type is auto and the fix is trivial.
An in-progress clang-tidy check
(https://github.com/llvm/llvm-project/pull/157213) attempts to find
instances where users copy a loop variable and then modify it.  In the
worst case, this pattern can cause bugs (because the user believes they
are modifying the value rather than a copy).  In cases where it's not a
bug, it can be considered a style improvement to default to using const
auto& declarations for clarity.

This CL fixes all instances of this trigger in non-test,
non-test-infrastructure, non-Blink, non-V8, non-third-party code in
Chromium where the fix is trivial (simply changing the loop declaration)
and the type is auto.  CLs for non-trivial fixes and non-auto
types will be made separately.

Bug: 424497405
Change-Id: I17bdcae733a3d2af809cb7a4bc84a7560c0164f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6986114
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Owners-Override: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522649}