Fix unintentional prerender aborts when committing the navigation in a reused FrameTree
The CL fixes the prerender aborts when reusing the FrameTree for
PrerenderHost. If the loading is still ongoing in the previous page when
the new prerender page commits, the previous load will fail with ABORT
error code. This will incorrectly cause the prerender of the new page to
fail before the fix.

Bug: 447995870
Change-Id: Ifb8f2f2b7627fecdd29cb5f4fa94c7941d2ed7c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6993101
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Keita Suzuki <suzukikeita@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Jiacheng Guo <gjc@google.com>
Cr-Commit-Position: refs/heads/main@{#1522650}