[InputVizard] Skip pointer transfer logic when Android fix is available.
In https://crbug.com/395596230, we found out new pointers on Chrome
window could go to Browser window, even when there's an already active
touch sequence on Viz. This has been since fixed as part of this
Google-internal Android bug: http://b/397979572.

In this change we are skipping this mitigation logic we added on Google
Android 16+ devices, since they shuould already receive pointer downs
directly on Viz. The fix is not available on all Android 16+ releases
since the fix was done late in Android 16 release cycle, and some of the
OEMs might not have picked up the fix.

Bug: 398208297
Change-Id: I7e7d1c75badbe88360e3238e8c1f50f4445c6e07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6994549
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522713}