Fix composedPath() to return at least currentTarget while dispatching
Chrome’s event path is implemented only in terms of node context.
Therefore, for an EventTarget, the event path is always empty.
However, while dispatching, the event path is populated, and
composedPath()[1] should return at least the currentTarget.
So, this patch fixes the bug by adding a check if the event is being
dispatched when the target is an EventTarget.

[1] https://dom.spec.whatwg.org/#dom-event-composedpath

Bug: 346766786
Change-Id: Icf139e2f4d709052550b3fa13685b40f4b8aede7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6984741
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Jinho Bang <zino@chromium.org>
Commit-Queue: Jinho Bang <zino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1522784}