[Settings/Autofill] Add password and passkey counts to Your Saved Info page
This change introduces a new C++ handler to fetch the number of saved
passwords and passkeys from Password Manager. This is part of the Home
of Transactions project, which aims to create a central "Your saved
info" page in settings.

The new handler is used by the TypeScript side to retrieve the counts on
the settings page. This avoids a dependency on the Passwords Private API
from the main settings UI, as a more performant and independent approach
to only retrieve the counts.

The counts are updated in real-time to reflect the changes occurring in
the password store from other tabs or sync.

This change only makes these counters available on 'Your saved info'
page as properties. They will be used and displayed once the data chips
components are ready.

Bug: 433613178
Change-Id: Icd82095a0c106a35093ac8bb601860b83d8bb2e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6987061
Reviewed-by: Stanislav Mikheyev <mikhe@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Ireneusz Szulc <ireneuszszulc@google.com>
Cr-Commit-Position: refs/heads/main@{#1522785}